Organize node context menu view tests
