+++ packages/extension/tests/webview/graph/contextMenu/node/view/folderActions.test.tsx
@@ -9,7 +9,7 @@ import {
   screen,
   setupGraphContextMenuTest,
   waitFor,
-} from './viewHarness';
+} from './harness';
 
 describe('Graph node context menu folder actions', () => {
   setupGraphContextMenuTest();
