+++ packages/extension/tests/webview/graph/contextMenu/node/view/symbolActions.test.tsx
@@ -9,7 +9,7 @@ import {
   setupGraphContextMenuTest,
   symbolData,
   waitFor,
-} from './viewHarness';
+} from './harness';
 
 describe('Graph node context menu symbol actions', () => {
   setupGraphContextMenuTest();
